requirements.txt
Django==1.6
Facebook==0.0
Pillow==2.2.1
boto==2.19.0
simplejson==3.3.1
